moved files and restored to discord login
