+++ src/app/testpage/page.tsx
@@ -1,5 +1,5 @@
-import { CreateFood } from '../_components/create-food';
-import { FullMenu } from '../_components/full-menu';
+import { CreateFood } from '../_components/createFood/create-food';
+import { FullMenu } from '../_components/fullMenu/full-menu';
 
 export default async function Testpage() {
   // const createFood = api.example.create.useMutation();
